+++ R/07-hdnom-kmplot.R
@@ -37,21 +37,21 @@
 #' time_new = smart$TEVENT[1001:2000]
 #' event_new = smart$EVENT[1001:2000]
 #'
-#' # Fit Cox model by lasso penalization
-#' lassofit = hdcox.lasso(x, Surv(time, event), nfolds = 5, rule = "lambda.1se", seed = 11)
+#' # Fit Cox model with lasso penalty
+#' fit = hdcox.lasso(x, Surv(time, event), nfolds = 5, rule = "lambda.1se", seed = 11)
 #'
-#' ### Internal calibration
+#' # Internal calibration
 #' cal.int = hdnom.calibrate(x, time, event, model.type = "lasso",
-#'                           alpha = 1, lambda = lassofit$'lasso_best_lambda',
+#'                           alpha = 1, lambda = fit$lasso_best_lambda,
 #'                           method = "cv", nfolds = 5,
 #'                           pred.at = 365 * 9, ngroup = 3)
 #'
 #' hdnom.kmplot(cal.int, group.name = c('High risk', 'Medium risk', 'Low risk'),
 #'              time.at = 1:6 * 365)
 #'
-#' ### External calibration
+#' # External calibration
 #' cal.ext =
-#'   hdnom.external.calibrate(lassofit, x, time, event,
+#'   hdnom.external.calibrate(fit, x, time, event,
 #'                            x_new, time_new, event_new,
 #'                            pred.at = 365 * 5, ngroup = 3)
 #'